Update lemonade_change.java
Lemonade Change is an easy greedy problem where each lemonade costs $5 and customers pay using $5, $10, or $20 bills. The goal is to give correct change to every customer in order, starting with no money. We maintain counts of available $5 and $10 bills and process customers one by one. For a $10 bill, we return one $5, and for a $20 bill, we greedily return a $10 and a $5 if possible, otherwise three $5 bills. If change cannot be given at any point, we return false. The algorithm runs in O(n) time and uses O(1) extra space.